requirements.txt
torch
Pillow
numpy 